scripts/requirements.txt
PyPDF2==3.0.1
Pillow==10.1.0
pytesseract==0.3.10
pymongo==4.6.1
python-dotenv==1.0.0
